// src/types/type.ts
export interface InitialState {
  q: string | undefined,
  minIngredients: string | undefined,
  maxIngredients: string | undefined,
  diet: string | undefined,
  health: string | undefined,
  cuisineType: string | undefined,
  mealType: string | undefined,
  dishType: string | undefined,
  minCalories: string | undefined,
  maxCalories: string | undefined,
  minTime: string | undefined,
  maxTime: string | undefined
}